refactor action sheet move to AddTags screen
add next and prev buttons on swiper
